Update is-porn.js
closes #3 

The issue is due to the use of the test() method on a RegExp object with the g (global) flag. When the test() method is used with a global regular expression, the lastIndex property of the RegExp object is automatically updated to point to the position in the string where the next match will start. This means that subsequent calls to test() will start from this position, leading to unexpected results.

It is also possible to fix this
```
const isOpenDNS = function(ip) {
                return OPENDNS_IPS.some(function(regexp) {
                    debug('trying', regexp);
                    regexp.lastIndex = 0; // Reset lastIndex
                    return regexp.test(ip);
                });
            };
```